+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### Changes on `master`
+
+ * Adds a scope `without_snaps_tag` to the revisioned models. Combinable with the existing`with_snaps_tag` scope.
+
+
 ## Version 0.0.1
 
 2014-03-26